Rename yaml util file
